Add a fuzzer for sparse bit set encoding.
