+++ fuzz/Cargo.toml
@@ -57,3 +57,9 @@ name = "fuzz_sparse_bit_set_decode"
 path = "fuzz_targets/fuzz_sparse_bit_set_decode.rs"
 test = false
 doc = false
+
+[[bin]]
+name = "fuzz_sparse_bit_set_encode"
+path = "fuzz_targets/fuzz_sparse_bit_set_encode.rs"
+test = false
+doc = false
